pyproject.toml
[project]
name = "kingsguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "causal-learn>=0.1.4.5",
    "crewai>=1.14.2",
    "crewai-tools>=1.14.2",
    "datasets==3.6.0",
    "docker>=7.1.0",
    "groq>=0.37.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "litellm>=1.83.0",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "peft>=0.19.1",
    "protobuf>=5.29.6",
    "pyarrow>=24.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "sentencepiece>=0.2.1",
    "torch>=2.11.0",
    "transformers>=5.6.2",
]
